% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addCombinations.R
\name{addCombinations}
\alias{addCombinations}
\title{Add human interpretable names to combinatorial states}
\usage{
addCombinations(multi.hmm, tracknames)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} object.}

\item{tracknames}{A vector of track names from which the combinations are generated. This vector must have the same length as the number of tracks in the \code{multi.hmm}. Identical entries will be treated as replicates.}
}
\value{
A \code{\link{chromstaR_multivariateHMM}} object with column 'combinations' added to \code{$bins} and \code{$segments}.
}
\description{
This function takes a \code{\link{chromstaR_multivariateHMM}} object and a vector of track names as input and translates the decimal combinatorial states into a human interpretable combination of track names.
}
\author{
Aaron Taudt
}

