% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustSensitivity.R
\name{adjustSensitivity}
\alias{adjustSensitivity}
\title{Adjust sensitivity of peak detection}
\usage{
adjustSensitivity(model, sensitivity = 0.01, invert = FALSE)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} or \code{\link{multiHMM}} object with posteriors.}

\item{sensitivity}{A vector of values between 0 and 1 for each column in \code{model$bins$posteriors}. If only one value is given, it will be reused for all columns. Values close to 0 will yield more stringent peak calls with lower false positive but higher false negative rate (i.e. more precise but less sensitive).}

\item{invert}{Select peaks below (\code{FALSE}) or above (\code{TRUE}) the given \code{sensitivity}. This is useful to select low confidence peaks.}
}
\value{
The input object is returned with adjusted peak calls.
}
\description{
Adjusts the peak calls of a \code{\link{uniHMM}}, \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object with a cutoff on the maximum-posterior within each peak. Lower values of \code{sensitivity} mean less sensitive and more precise peak calls. Remaining peaks are kept intact, as opposed to function \code{\link{changePostCutoff}}, where broad peaks are fragmented.
}
\details{
Each peak has a maximum-posterior (maxPostInPeak, between 0 and 1) associated. The sensitivity is adjusted with a simple cutoff on 1-maxPostInPeak, e.g. for \code{sensitivity = 0.01} only peaks with a maximum-posterior \code{>= 0.99} will be selected.
}
\examples{
## Get an example BAM file
file <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bam",
                       package="chromstaRData")
## Bin the file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(file, assembly=rn4_chrominfo, binsizes=1000,
                  stepsizes=500, chromosomes='chr12')
## Fit the univariate Hidden Markov Model
hmm <- callPeaksUnivariate(binned, max.time=60, eps=1)
## Compare fits with different fdrs
plotHistogram(hmm) + ylim(0,0.25) + ylim(0,0.3)
plotHistogram(adjustSensitivity(hmm, sensitivity=0.01)) + ylim(0,0.3)
plotHistogram(adjustSensitivity(hmm, sensitivity=1e-12)) + ylim(0,0.3)

}
\seealso{
\code{\link{changePostCutoff}}
}
\author{
Aaron Taudt
}
