% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{align2binned}
\alias{align2binned}
\title{Convert aligned reads from various file formats into read counts in equidistant bins}
\usage{
align2binned(file, format, assembly, bamindex = file,
  pairedEndReads = FALSE, chrom.length.file = NULL,
  outputfolder = "binned_data", binsizes = 500, chromosomes = NULL,
  save.as.RData = FALSE, min.mapq = 10, downsample.to.reads = NULL,
  remove.duplicate.reads = FALSE, max.fragment.width = 1000)
}
\arguments{
\item{file}{A file with aligned reads.}

\item{format}{One of \code{c('bam', 'bed', 'bedGraph')}.}

\item{assembly}{An assembly to specify the chromosome lengths. One of \code{c('hg19','hg18')}.}

\item{bamindex}{Index file if \code{format='bam'} with or without the .bai ending. If this file does not exist it will be created and a warning is issued.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{chrom.length.file}{A file which contains the chromosome lengths in basepairs. The first column contains the chromosome name and the second column the length (see also \code{\link{chrom.length.file}}.}

\item{outputfolder}{Folder to which the binned data will be saved if \code{save.as.RData=TRUE}. If the specified folder does not exist, it will be created.}

\item{binsizes}{A vector with integer values which will be used for the binning. If more than one value is given, output files will be produced for each bin size.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{save.as.RData}{If set to \code{TRUE}, output will be written to file. The filename is generated from the input file and bin sizes. If set to \code{FALSE}, no output file will be written. Instead, a \link{GenomicRanges} object containing the binned data will be returned. Only the first binsize will be processed in this case.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{downsample.to.reads}{Downsample the input data to the specified number of reads.}

\item{remove.duplicate.reads}{Logical indicating whether duplicate reads should be removed. NOTE: Duplicate removal only works if your duplicate reads are properly flagged in the BAM file.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments due to mapping errors of paired end reads.}
}
\value{
The function produces a \code{\link{GRanges}} object with one meta data column 'reads' that contains the read count. This binned data will be either written to file (\code{save.as.RData=FALSE}) or given as return value (\code{save.as.RData=FALSE}).
}
\description{
Convert aligned reads in .bam or .bed format into read counts in equidistant windows. Convert signal values in .bedGraph format to signal counts in equidistant windows.
}

