% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binReads.R
\name{blacklistGRanges}
\alias{blacklistGRanges}
\title{Remove blacklist from GRanges}
\usage{
blacklistGRanges(gr, blacklist = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{blacklist}{Filename of the bed or bed.gz file which contains the blacklisted regions.}
}
\value{
A \code{\link{GRanges}} object with the regions not overlapping any regions of the blacklist.
}
\description{
This is a simple convenience function to remove regions from a bed(.gz)-file from a \code{\link{GRanges}} object. The bed-file is expected to have at least the following three fields: \code{chromosome, start, end}.
}
\author{
Aaron Taudt, David Widmann
}

