% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callPeaksReplicates.R
\name{callPeaksReplicates}
\alias{callPeaksReplicates}
\title{Fit a multivariate Hidden Markov Model to multiple ChIP-seq replicates}
\usage{
callPeaksReplicates(hmm.list, max.states = 32, force.equal = FALSE,
  eps = 0.01, max.iter = NULL, max.time = NULL, num.threads = 1,
  max.distance = 0.2)
}
\arguments{
\item{hmm.list}{A list of \code{\link{chromstaR_univariateHMM}}s generated by \code{\link{callPeaksUnivariate}}, e.g. \code{list(hmm1,hmm2,...)} or \code{c("file1","file2",...)}. Alternatively, this parameter also accepts a \code{\link{chromstaR_multivariateHMM}} and will check if the distance between replicates is greater than \code{max.distance}.}

\item{max.states}{The maximum number of combinatorial states to consider. The default (32) is sufficient to treat up to 5 replicates exactly and more than 5 replicates approximately.}

\item{force.equal}{The default (\code{FALSE}) allows replicates to differ in their peak-calls, although the majority will usually be identical. If \code{force.equal=TRUE}, all peaks will be identical among all replicates.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.iter}{The maximum number of iterations for the Baum-Welch algorithm. The default \code{NULL} is no limit.}

\item{max.time}{The maximum running time in seconds for the Baum-Welch algorithm. If this time is reached, the Baum-Welch will terminate after the current iteration finishes. The default \code{NULL} is no limit.}

\item{num.threads}{Number of threads to use. Setting this to >1 may give increased performance.}

\item{max.distance}{This number is used as a cutoff to group replicates based on their distance matrix. The lower this number, the more similar replicates have to be to be grouped together.}
}
\value{
Output is a \code{\link{chromstaR_multivariateHMM}} object with additional entry \code{replicateInfo}. If only one \code{\link{chromstaR_univariateHMM}} was given as input, a simple list() with the \code{replicateInfo} is returned.
}
\description{
Fit an HMM to multiple ChIP-seq replicates and derive correlation measures. Input is a list of \code{\link{chromstaR_univariateHMM}}s generated by \code{\link{callPeaksUnivariate}}.
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{chromstaR_multivariateHMM}}, \code{\link{callPeaksUnivariate}}, \code{\link{callPeaksMultivariate}}
}

