% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changePostCutoff.R
\name{changePostCutoff}
\alias{changePostCutoff}
\title{Change the posterior cutoff of a Hidden Markov Model}
\usage{
changePostCutoff(model, post.cutoff = 0.5)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} or \code{\link{multiHMM}} object with posteriors.}

\item{post.cutoff}{A vector of posterior cutoff values between 0 and 1 the same length as \code{ncol(model$bins$posteriors)}. If only one value is given, it will be reused for all columns. Values close to 1 will yield more stringent peak calls with lower false positive but higher false negative rate.}
}
\value{
The input object is returned with adjusted peak calls.
}
\description{
Adjusts the peak calls of a \code{\link{uniHMM}} or \code{\link{multiHMM}} object with the given posterior cutoff.
}
\details{
Posterior probabilities are between 0 and 1. Peaks are called if the posteriors for a state (univariate) or sample (multivariate) are >= \code{post.cutoff}.
}
\author{
Aaron Taudt
}

