% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{chromstaR-objects}
\alias{chromstaR-objects}
\title{chromstaR objects}
\description{
\pkg{\link{chromstaR}} defines several objects.
\itemize{
\item \code{\link{uniHMM}}: Returned by \code{\link{callPeaksUnivariate}}.
\item \code{\link{multiHMM}}: Returned by \code{\link{callPeaksMultivariate}} and \code{\link{callPeaksReplicates}}.
\item \code{\link{combinedMultiHMM}}: Returned by \code{\link{combineMultivariates}}.
}
}
