% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromstaR-package.R
\docType{package}
\name{chromstaR-package}
\alias{chromstaR}
\alias{chromstaR-package}
\title{Combinatorial and differential chromatin state analysis for ChIP-seq data}
\description{
This package implements functions for the combinatorial and differential analysis of ChIP-seq data. It was developed for histone modifications with a broad profile but is also suitable for the analysis of transcription factor binding data. A Hidden Markov Model with a mixture of Negative Binomials as emission densities is used to call peaks. Please read our publication for a detailed description of the method. TODO: insert publication.
}
\details{
The main function of this package is \code{\link{Chromstar}}. For a detailed introduction type \code{browseVignettes("chromstaR")} and read the vignette.
}
\author{
Aaron Taudt, Maria Colome-Tatche, Matthias Heinig, Minh Anh Nguyen
}

