% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromstaR.R
\docType{package}
\name{chromstaR}
\alias{chromstaR}
\alias{chromstaR-package}
\title{Combinatorial analysis of ChIP-seq histone modification data using Hidden Markov Models}
\description{
\pkg{chromstaR} is an R package for the combinatorial or differential analysis of multiple ChIP-seq experiments. It was developed for histone modifications with a broad profile but is also suitable for the analysis of transcription factor binding data. A Hidden Markov Model with a mixture of Negative Binomials as emission densities is used to call peaks. Please read our publication for a detailed description of the method. TODO: insert publication
}
\author{
Aaron Taudt, Maria Colome Tatche, Matthias Heinig
}

