% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combinatorialEntropy.R
\name{combinatorialEntropy}
\alias{combinatorialEntropy}
\title{Combinatorial entropy}
\usage{
combinatorialEntropy(multi.hmm.list, window.size.bp = NULL,
  direction.of.adding = "between.hmms")
}
\arguments{
\item{multi.hmm.list}{A list of \code{\link{chromstaR_multivariateHMM}} objects or a vector of files that contain such objects.}

\item{window.size.bp}{Window size in base-pairs that will be used to average the results.}

\item{direction.of.adding}{Calculation of the combinatorial entropy only makes sense between different modifications. Choose 'between.hmms' if you have one modification per HMM, which is the case in a differential analysis. Choose 'inside.hmms' if you have multiple modifications per HMM, which is the case in a combinatorial analysis. In this case it is up to the user to make sure that the \code{IDs} between HMMs are identical.}
}
\description{
Calculate the combinatorial entropy along the genome for a set of \code{\link{chromstaR_multivariateHMM}}s.
}
\details{
The combinatorial entropy describes the variation between samples. For each genomic position, it is defined as
\deqn{ log( choose(n,x) ) }
where \eqn{n} is the number of samples and \eqn{x} the number of samples that are in state 'modified'. Calculation of this entropy measure only makes sense for samples with the same modification (e.g. 7 samples of H3K36me3 in different tissues). For multiple modifications, the entropy can be calculated for each modification separately and then added.
}
\author{
Aaron Taudt
}

