% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{combinatorialFrequency}
\alias{combinatorialFrequency}
\title{Frequency of combinatorial states}
\usage{
combinatorialFrequency(multi.hmm)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} or a file that contains such an object.}
}
\value{
A table with frequencies of each combinatorial state.
}
\description{
Get the genomewide frequency of each combinatorial state.
}
\author{
Aaron Taudt
}

