% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{differentialScoreMax}
\alias{differentialScoreMax}
\title{Calculate maximum differential score}
\usage{
differentialScoreMax(posteriors, info)
}
\arguments{
\item{posteriors}{A matrix with posterior probabilities.}

\item{info}{An \code{\link{experiment.table}} with additional column 'ID'.}
}
\value{
A vector with differential score.
}
\description{
Calculate maximum differential score from posteriors.
}

