% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{differentialScoreSum}
\alias{differentialScoreSum}
\title{Calculate added differential score}
\usage{
differentialScoreSum(posteriors, info)
}
\arguments{
\item{posteriors}{A matrix with posterior probabilities.}

\item{info}{An \code{\link{experiment.table}} with additional column 'ID'.}
}
\value{
A vector with differential score.
}
\description{
Calculate added differential score from posteriors.
}

