% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichmentCurve.R
\name{enrichmentCurve}
\alias{enrichmentCurve}
\title{Enrichment of (combinatorial) states for genomic annotations}
\usage{
enrichmentCurve(hmm, annotation, bp.around.annotation = 10000,
  region = c("start", "inside", "end"))
}
\arguments{
\item{hmm}{A \code{\link{chromstaR_multivariateHMM}} object.}

\item{annotation}{A \code{\link{GRanges}} object with the annotation of interest.}

\item{bp.around.annotation}{An integer specifying the number of basepairs up- and downstream of the annotation for which the enrichment will be calculated.}

\item{region}{A combination of \code{c('start','inside','end')} specifying the region of the annotation for which the enrichment will be calculated.}
}
\value{
A \code{list()} containing \code{data.frame()}s for enrichment of combinatorial states and binary states at the start, end and inside of the annotation.
}
\description{
The function calculates the enrichment of a genomic feature with peaks or combinatorial states. Input is a \code{\link{chromstaR_multivariateHMM}} object (containing the peak calls and combinatorial states) and a \code{\link{GRanges}} object containing the annotation of interest (e.g. transcription start sites or genes).
}
\author{
Aaron Taudt
}

