\name{example.multi.HMM}
\docType{data}
\alias{example.multi.HMM}
\title{Multivariate Hidden Markov Model for usage in examples}
\description{A \code{\link{chromstaR_multivariateHMM}} object, generated from the brain tissue dataset (H3K36me3, 7 tissues) in 'extdata/brain/'. This dataset is used in the example section of \pkg{chromstaR} functions.}
\usage{
data(example.multi.HMM)
}
\format{
See \code{\link{chromstaR_multivariateHMM}} for a description of the format.
}

\keyword{example}
\keyword{multivariate}
