% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{experiment.table}
\alias{experiment.table}
\title{Experiment data table}
\format{A \code{data.frame} with columns 'file', 'mark', 'condition', 'replicate' and 'pairedEndReads'.}
\description{
A \code{data.frame} specifying the structure of the experiment.
}
\examples{
data(experiment_table)
print(experiment_table)

}

