% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedimport.R
\name{exportBed3File}
\alias{exportBed3File}
\title{Write GRanges to bed3-file}
\usage{
exportBed3File(gr, filename, append = FALSE, chromosome.format = "NCBI")
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{filename}{The name of the file that will be written. The ending ".bed.gz". Any existing file will be overwritten.}

\item{append}{Whether or not to append to an existing file.}

\item{chromosome.format}{Desired format of the chromosomes. Either 'NCBI' for (1,2,3 ...) or 'UCSC' for (chr1,chr2,chr3 ...).}
}
\value{
\code{NULL}
}
\description{
This is a simple convenience function to write a \code{\link{GRanges}} object to a bed(.gz)-file. The bed-file will have the following three fields: \code{chromosome, start, end}.
}
\author{
Aaron Taudt, David Widmann
}

