% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportBinnedData}
\alias{exportBinnedData}
\title{Export genome browser viewable files}
\usage{
exportBinnedData(binned.data.list, filename, header = TRUE,
  separate.files = TRUE)
}
\arguments{
\item{binned.data.list}{A \code{list()} of \code{\link{binned.data}} objects or vector of files that contain such objects.}

\item{filename}{The name of the file that will be written. The ending ".wig.gz" for read counts will be appended. Any existing file will be overwritten.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for each object in \code{binned.data.list}.}
}
\value{
\code{NULL}
}
\description{
Export read counts (RPKM) as genome browser viewable file
}
\details{
Export read counts from \code{\link{binned.data}} as a file which can be uploaded into a genome browser. Read counts are exported in WIGGLE format as RPKM (.wig.gz).
}
\examples{
## Get an example BAM file
file <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bam",
                       package="chromstaRData")
## Bin the file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(file, assembly=rn4_chrominfo, binsizes=1000,
                  chromosomes='chr12')
## Export the binned read counts
exportBinnedData(list(binned), filename=tempfile())

}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportMultivariate}}
}

