% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFiles.R
\name{exportFiles}
\alias{exportCombinations}
\alias{exportCounts}
\alias{exportFiles}
\alias{exportPeaks}
\title{Export genome browser uploadable files}
\usage{
exportPeaks(model, filename, header = TRUE, separate.files = TRUE,
  trackname = NULL)

exportCounts(model, filename, header = TRUE, separate.files = TRUE,
  trackname = NULL)

exportCombinations(model, filename, header = TRUE, separate.files = TRUE,
  trackname = NULL, exclude.states = "[]", include.states = NULL)
}
\arguments{
\item{model}{A \code{\link{chromstaR-objects}}.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either "_peaks.bed.gz" for peak-calls or "_counts.wig.gz" for read counts or "_combinations.bed.gz" for combinatorial states. Any existing file will be overwritten.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for each track.}

\item{trackname}{Name that will be used in the "track name" field of the BED file.}

\item{exclude.states}{A character vector with combinatorial states that will be excluded from export.}

\item{include.states}{A character vector with combinatorial states that will be exported. If specified, \code{exclude.states} is ignored.}
}
\value{
\code{NULL}
}
\description{
These functions allow to export \code{\link{chromstaR-objects}} as files which can be uploaded to a genome browser. Peak calls are exported in BED format (.bed.gz), read counts in wiggle format (.wig.gz) as RPKM values, and combinatorial states are exported in BED format (.bed.gz).
}
\section{Functions}{
\itemize{
\item \code{exportPeaks}: Export peak calls in BED format.

\item \code{exportCounts}: Export read counts as RPKM values in wiggle format.

\item \code{exportCombinations}: Export combinatorial states in BED format.
}}
\examples{
## Get an example multiHMM
file <- system.file("data","combined_mode-differential.RData",
                    package="chromstaR")
model <- get(load(file))
## Export peak calls and combinatorial states
exportPeaks(model, filename=tempfile())
exportCombinations(model, filename=tempfile())

}

