% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.R
\name{exportUnivariates}
\alias{exportUnivariates}
\title{Export genome browser viewable files}
\usage{
exportUnivariates(hmm.list, filename, what = c("peaks", "reads"),
  header = TRUE, separate.files = FALSE, orderByScore = TRUE)
}
\arguments{
\item{hmm.list}{A list of \code{\link{chromstaR_univariateHMM}} objects or files that contain such objects.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either ".bed.gz" for peak-calls or ".wig.gz" for read counts. Any existing file will be overwritten.}

\item{what}{A character vector specifying what will be exported. Supported are \code{c('peaks', 'reads')}.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for each hmm in \code{hmm.list}.}

\item{orderByScore}{Logical indicating whether or not to sort entries in BED file by score.}
}
\description{
Export univariate peak-calls and read counts as genome browser viewable file
}
\details{
Export \code{\link{chromstaR_univariateHMM}} objects as files which can be uploaded into a genome browser. Peak-calls are exported in BED format (.bed.gz) and read counts are exported in WIGGLE format (.wig.gz).
}
\examples{
### Univariate peak calling ###
## Get example BED-files with ChIP-seq reads for H3K36me3
# in 7 different brain tissues (chr22)
bedfiles <- list.files(system.file(file.path("extdata","brain"),
                      package="chromstaR"), full=TRUE)
## Bin the data into bin size 1000bp and build the univariate Hidden Markov Model (HMM)
binned.data.list <- list()
uni.HMMs <- list()
for (bedfile in bedfiles) {
 binned.data <- bed2binned(bedfile, assembly='hg19', binsize=1000, save.as.RData=FALSE)
 binned.data.list[[bedfile]] <- binned.data
 uni.HMMs[[bedfile]] <- callPeaksUnivariate(binned.data, ID=basename(bedfile),
                                            max.time=30, eps=0.01)
}
## Export the binned read counts and peaks
\donttest{exportUnivariates(uni.HMMs, filename='chromstaR-example_univariate.HMMs', what=c('reads','peaks'))}
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportBinnedData}}, \code{\link{exportMultivariate}}
}

