% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{foldEnrichment}
\alias{foldEnrichment}
\title{Fold enrichment of combinatorial states}
\usage{
foldEnrichment(multi.hmm, featurelist)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} or a file that contains such an object.}

\item{featurelist}{A list with \code{\link{Granges}} objects containing coordinates of multiple features. The names of the list entries will be used to name the return values.}

\item{feature}{A \code{\link{GRanges}} with coordinates of the feature to compute the fold enrichment to.}
}
\value{
A named array with fold enrichments.
}
\description{
Compute the fold enrichment of combinatorial states in a given feature (e.g. TSS, exons, ...)
}
\author{
Aaron Taudt
}

