% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{genes_rn4}
\alias{genes_rn4}
\title{Gene coordinates for rn4}
\format{A data.frame.}
\description{
A data.frame containing gene coordinates and biotypes of the rn4 assembly.
}
\examples{
data(genes_rn4)
head(genes_rn4)
}
