% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{genomicFrequencies}
\alias{genomicFrequencies}
\title{Frequencies of combinatorial states}
\usage{
genomicFrequencies(multi.hmm, combinations = NULL)
}
\arguments{
\item{multi.hmm}{A \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object or a file that contains such an object.}

\item{combinations}{A vector with combinations for which the frequency will be calculated. If \code{NULL} all combinations will be considered.}
}
\value{
A table with frequencies of each combinatorial state.
}
\description{
Get the genomewide frequency of each combinatorial state.
}
\examples{
## Prepare the file paths. Exchange this with your input and output directories.
inputfolder <- system.file("extdata","euratrans", package="chromstaRData")
outputfolder <- file.path(tempdir(), 'liver-example')
## Define experiment structure
data(experiment_table_liver)
## Define assembly
# This is only necessary if you have BED files, BAM files are handled automatically.
# For common assemblies you can also specify them as 'hg19' for example.
data(rn4_chrominfo)
## Run ChromstaR
Chromstar(inputfolder, experiment.table=experiment_table_liver,
         outputfolder=outputfolder, numCPU=2, binsize=1000, assembly=rn4_chrominfo,
         prefit.on.chr='chr12', mode='mark', eps=1)
## Load a multivariate file and get frequencies
multimodel <- file.path(outputfolder,'multivariate',
                       'multivariate_mode-mark_condition-liver.RData')
genomicFrequencies(multimodel)
}
\author{
Aaron Taudt
}

