% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDistinctColors.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{Get distinct colors}
\usage{
getDistinctColors(n, exclude.colors = c("white", "black", "gray", "grey"),
  start.color = "blue")
}
\arguments{
\item{n}{Number of colors to select.}

\item{exclude.colors}{Character vector with colors that should not be used.}

\item{start.color}{Color to start the selection process from.}
}
\value{
A character vector with colors.
}
\description{
Get a set of distinct colors selected from \code{\link{colors}}.
}
\details{
The function computes the euclidian distance between all \code{\link{colors}} and iteratively selects those that have the furthest closes distance to the set of already selected colors.
}
\author{
Aaron Taudt
}

