% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getStateColors}
\alias{getStateColors}
\title{Get state colors}
\usage{
getStateColors(labels)
}
\arguments{
\item{labels}{Any combination of \code{c("zero-inflation","unmodified","modified", "total", "counts")}.}
}
\value{
A character vector with colors.
}
\description{
Get the colors that are used for plotting.
}

