% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapCombinations}
\alias{heatmapCombinations}
\title{Plot a heatmap of combinatorial states}
\usage{
heatmapCombinations(hmm = NULL, marks = NULL, vertical = FALSE)
}
\arguments{
\item{hmm}{A \code{\link{multiHMM}} object.}

\item{marks}{A character vector with histone marks. If specified, \code{hmm} will be ignored.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object containing the plot.
}
\description{
Plot a heatmap that shows the binary presence/absence of marks for the different combinations.
}
\examples{
marks <- c('H3K4me3','H3K27me3','H4K20me1')
heatmapCombinations(marks=marks, vertical=FALSE)

}
\author{
Aaron Taudt
}

