% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapCountCorrelation}
\alias{heatmapCountCorrelation}
\title{Read count correlation heatmap}
\usage{
heatmapCountCorrelation(model, cluster = TRUE)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object or file that contains such an object.}

\item{cluster}{Logical indicating whether or not to cluster the heatmap.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Heatmap of read count correlations (see \code{\link[stats]{cor}}).
}
\seealso{
\code{\link{plotting}}
}

