% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapTiestrengths}
\alias{heatmapTiestrengths}
\title{Heatmap of tie strengths}
\usage{
heatmapTiestrengths(model, normalize.tiestrengths = FALSE)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{normalize.tiestrengths}{Set to \code{TRUE} to normalize tiestrengths per column instead of per row. This is useful to see which tracks are influenced by which other tracks.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a heatmap of tie strengths for a \code{\link{multiHMM}} model.
}
\examples{
## Get an example multiHMM ##
file <- system.file("data","multivariate_mode-combinatorial_condition-SHR.RData",
                    package="chromstaR")
model <- get(load(file))
## Plot transition probabilites as heatmap
heatmapTiestrengths(model)

}
\seealso{
\code{\link{plotting}}
}
