% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOtherSegmentations.R
\name{importChromHMMsegmentation}
\alias{importChromHMMsegmentation}
\title{Import ChromHMM segmentation}
\usage{
importChromHMMsegmentation(inputfolder, outputfolder, assembly, binsize,
  experiment.table, chrom.lengths = NULL, chromosomes = NULL,
  chromosome.format = "UCSC")
}
\arguments{
\item{inputfolder}{Folder with ChromHMM output files (segmentation, emission and transition probabilities).}

\item{outputfolder}{Output folder for the converted \code{\link{chromstaR-objects}}.}

\item{assembly}{An assembly from which the chromosome lengths are determined. Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}.}

\item{binsize}{The bin size in base pairs used in ChromHMM.}

\item{experiment.table}{A \code{data.frame} or tab-separated text file with the structure of the experiment. See \code{\link{experiment.table}} for an example.}

\item{chrom.lengths}{A named character vector with chromosome lengths. Names correspond to chromosomes.}

\item{chromosomes}{A subset of chromosomes for which the bins are generated.}

\item{chromosome.format}{A character specifying the format of the chromosomes if \code{assembly} is specified. Either 'NCBI' for (1,2,3 ...) or 'UCSC' for (chr1,chr2,chr3 ...).}
}
\value{
A \code{\link{combinedMultiHMM}} object. Intermediate \code{\link{multiHMM}} objects are saved to \code{outputfolder}.
}
\description{
Import a chromatin state segmentation produced by ChromHMM as \code{\link{chromstaR-objects}}. This is useful to perform comparative analyses or simply use chromstaR plotting or enrichment functions on a ChromHMM segmentation.
}
\details{
The function takes the *segments.bed files, emissions_*.txt and transitions_*.txt files in \code{inputfolder}, and converts them into \code{\link{multiHMM}} objects for each condition that was specified in the \code{experiment.table}. All \code{\link{multiHMM}} objects are then combined into a \code{\link{combinedMultiHMM}} object and returned.
}
\examples{
inputfolder <- "differential_CD4_numstates_16_binsize_1000bp"
outputfolder <- "testconvertChromHMM2chromstar"
assembly <- "mm9"
binsize <- 1000
chrom.lengths <- NULL
chromosomes <- paste0("chr", c(1:19, "X"))
chromosome.format <- "UCSC"
experiment.table <- "differential_CD4_numstates_16_binsize_1000bp/experiment_table_differential_CD4.tsv"
model <- importChromHMMsegmentation(inputfolder, outputfolder, assembly, binsize, experiment.table, chrom.lengths, chromosomes, chromosome.format)

### Plot transition and emission probabilities of ChromHMM model ###
heatmapTransitionProbs(reorder.states=FALSE, transitionProbs=model$transitionProbs)
heatmapCombinations(emissionProbs = model$emissionProbs)

### Use plotEnrichment function on imported ChromHMM segmentation ###
library(biomaRt)
ensembl <- useMart('ENSEMBL_MART_ENSEMBL', host='may2012.archive.ensembl.org',
                  dataset='mmusculus_gene_ensembl')
genes <- getBM(attributes=c('ensembl_gene_id', 'chromosome_name', 'start_position',
                           'end_position', 'strand', 'external_gene_id',
                           'gene_biotype'),
              mart=ensembl)
# Transform to GRanges for easier handling
genes <- GRanges(seqnames=paste0('chr',genes$chromosome_name),
                ranges=IRanges(start=genes$start, end=genes$end),
                strand=genes$strand,
                name=genes$external_gene_id, biotype=genes$gene_biotype)
print(genes)

ggplts <- plotEnrichment(model, annotation=genes)
ggplts[[1]] + facet_wrap(~combination)

}
