% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChroms.R
\name{mergeChroms}
\alias{mergeChroms}
\title{Merge several \code{\link{chromstaR_multivariateHMM}}s into one object}
\usage{
mergeChroms(multi.hmm.list, filename = NULL)
}
\arguments{
\item{multi.hmm.list}{A list of \code{\link{chromstaR_multivariateHMM}} objects.}

\item{filename}{The file name where the merged object will be stored. If \code{filename} is not specified, a \code{\link{chromstaR_multivariateHMM}} is returned.}
}
\value{
A \code{\link{chromstaR_multivariateHMM}} object or NULL, depending on option \code{filename}.
}
\description{
Merge several \code{\link{chromstaR_multivariateHMM}}s into one object. This can be done to merge fits for separate chromosomes into one object for easier handling. Merging will only be done if all models have the same IDs.
}
\author{
Aaron Taudt
}

