% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{model.combined}
\alias{model.combined}
\title{Combined multivariate HMM for demonstration purposes}
\format{A \code{\link{combinedMultiHMM}} object.}
\description{
A \code{\link{combinedMultiHMM}} object for demonstration purposes in examples of package \pkg{\link{chromstaR}}.
}
\examples{
## Get an example combinedMultiHMM
file <- system.file("data","combined_mode-differential.RData",
                    package="chromstaR")
model <- get(load(file))
}
