% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{model.univariate}
\alias{model.univariate}
\title{Univariate HMM for demonstration purposes}
\format{A \code{\link{uniHMM}} object.}
\description{
A \code{\link{uniHMM}} object for demonstration purposes in examples of package \pkg{\link{chromstaR}}.
}
\examples{
## Get an example uniHMM
file <- system.file("data","H3K27me3-BN-rep1.RData", package="chromstaR")
model <- get(load(file))
}
