% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{multiHMM}
\alias{multiHMM}
\alias{multi.hmm}
\title{Multivariate HMM object}
\value{
A \code{list()} with the following entries:
\item{info}{Experiment table for this object.}
\item{bincounts}{A \code{\link[GenomicRanges]{GRanges-class}} object containing the genomic bin coordinates and original binned read count values for different offsets.}
\item{bins}{A \code{\link[GenomicRanges]{GRanges-class}} object containing the genomic bin coordinates, their read count, (optional) posteriors and state classification.}
\item{segments}{Same as \code{bins}, but consecutive bins with the same state are collapsed into segments.}
\item{peaks}{A \code{list()} with \code{\link[GenomicRanges]{GRanges-class}} containing peak coordinates for each ID in \code{info}.}
\item{mapping}{A named vector giving the mapping from decimal combinatorial states to human readable combinations.}
\item{weights}{Weight for each component. Same as \code{apply(hmm$posteriors,2,mean)}.}
\item{weights.univariate}{Weights of the univariate HMMs.}
\item{transitionProbs}{Matrix of transition probabilities from each state (row) into each state (column).}
\item{transitionProbs.initial}{Initial \code{transitionProbs} at the beginning of the Baum-Welch.}
\item{startProbs}{Probabilities for the first bin. Same as \code{hmm$posteriors[1,]}.}
\item{startProbs.initial}{Initial \code{startProbs} at the beginning of the Baum-Welch.}
\item{distributions}{Emission distributions used for this model.}
\item{convergenceInfo}{Contains information about the convergence of the Baum-Welch algorithm.}
\item{convergenceInfo$eps}{Convergence threshold for the Baum-Welch.}
\item{convergenceInfo$loglik}{Final loglikelihood after the last iteration.}
\item{convergenceInfo$loglik.delta}{Change in loglikelihood after the last iteration (should be smaller than \code{eps})}
\item{convergenceInfo$num.iterations}{Number of iterations that the Baum-Welch needed to converge to the desired \code{eps}.}
\item{convergenceInfo$time.sec}{Time in seconds that the Baum-Welch needed to converge to the desired \code{eps}.}
\item{correlation.matrix}{Correlation matrix of transformed reads.}
}
\description{
The multivariate HMM object is output of the function \code{\link{callPeaksMultivariate}} and is a \code{list()} with various entries. The \code{class()} attribute of this list was set to "multiHMM". For a given hmm, the entries can be accessed with the list operators 'hmm[[]]' or 'hmm$'.
}
\examples{
## Get an example multiHMM
file <- system.file("data","multivariate_mode-combinatorial_condition-SHR.RData",
                    package="chromstaR")
model <- get(load(file))
}
\seealso{
\code{\link{callPeaksMultivariate}}, \code{\link{uniHMM}}, \code{\link{combinedMultiHMM}}
}
