% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{multimodel}
\alias{multimodel}
\title{Multivariate HMM for demonstration purposes}
\format{A \code{\link{multiHMM}} object.}
\description{
A \code{\link{multiHMM}} object for demonstration purposes in examples of package \pkg{\link{chromstaR}}.
}

