% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.character}
\alias{plot.character}
\title{Plotting function for saved \pkg{\link{chromstaR}} objects}
\usage{
\method{plot}{character}(x, ...)
}
\arguments{
\item{x}{A filename that contains either \code{\link{binned.data}}, a \code{\link{uniHMM}} or a \code{\link{multiHMM}}.}

\item{...}{Additional arguments.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Convenience function that loads and plots a \pkg{\link{chromstaR}} object in one step.
}

