% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.chromstaR_multivariateHMM}
\alias{plot.chromstaR_multivariateHMM}
\title{Plotting function for \code{\link{chromstaR_multivariateHMM}} objects}
\usage{
\method{plot}{chromstaR_multivariateHMM}(x, type = "transitionMatrix", ...)
}
\arguments{
\item{x}{A \code{\link{chromstaR_multivariateHMM}} object.}

\item{type}{Type of the plot, one of \code{c('transitionMatrix','histograms','correlation')}. You can also specify the type with an integer number.
\describe{
  \item{\code{transitionMatrix}}{A heatmap with entries of the transition matrix.}
  \item{\code{histograms}}{Fitted histograms of all underlying univariate distributions.}
}}

\item{...}{Additional arguments for the different plot types.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make different types of plots for \code{\link{chromstaR_multivariateHMM}} objects.
}

