% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.chromstaR_univariateHMM}
\alias{plot.chromstaR_univariateHMM}
\title{Plotting function for \code{\link{chromstaR_univariateHMM}} objects}
\usage{
\method{plot}{chromstaR_univariateHMM}(x, type = "histogram", ...)
}
\arguments{
\item{x}{A \code{\link{chromstaR_univariateHMM}} object.}

\item{type}{Type of the plot, one of \code{c('histogram', 'boxplot', 'normalTransformation')}. You can also specify the type with an integer number.
\describe{
  \item{\code{histogram}}{A histogram of binned read counts with fitted mixture distribution.}
  \item{\code{boxplot}}{A boxplot of read counts for the different states.}
  \item{\code{normalTransformation}}{A histogram of transformed read counts.}
}}

\item{...}{Additional arguments for the different plot types.
\describe{
  \item{\code{state}}{Plot the \code{histogram}, \code{boxplot} or \code{normalTransformation} only for the specified state. One of \code{c('unmodified','modified')}.}
  \item{\code{chromosome,start,end}}{Plot the \code{histogram} only for the specified chromosome, start and end position.}
}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make different types of plots for \code{\link{chromstaR_univariateHMM}} objects.
}

