\name{plot.distribution}
\alias{plot.distribution}
\title{
Plot the emission distributions that were fitted to the read data
}
\description{
Plot emission distribution of the input data and fitted Negative Binomial distributions of the HMM.
}
\usage{
plot.distribution(model, chr=NULL, start=NULL, end=NULL, return.plot=FALSE)
}
\arguments{
\item{model}{
A model generated by \code{\link{univariate.from.binned.data}}.
}
\item{chr}{
A string describing a chromosome, e.g. "chr1". This will plot only the histogram and fits for this chromosome.
}
\item{start}{
An integer specifying the start position of the region to plot. If chromosome is unset and there is more than one chromosome in the model, this option will be ignored.
}
\item{end}{
An integer specifying the end position of the region to plot. If chromosome is unset and there is more than one chromosome in the model, this option will be ignored.
}
\item{return.plot}{
If set to 'TRUE', a ggplot2 object containing the histogram with fits will be returned. You need to install the ggplot2 package to use this functionality. Plotting like this can take more time to render!
}
}

\details{
The function expects a model generated by \code{\link{univariate.from.binned.data}}. It will plot the histogram of the binned reads along with the fitted distributions of the univariate HMM.
}
\keyword{histogram}
\keyword{distributions}
\keyword{univariate}
\references{}
\author{Aaron Taudt}
\note{}
\seealso{}
\examples{
# Load example datasets with genome coordinates and number of aligned reads for each bin
data(binned.data.example.1)
# Do the univariate HMMs
model = univariate.from.binned.data(binned.data.example.1, eps=1, max.time=60)
# Plot histogram of binned reads and fitted distributions
plot.distribution(model)
}

