% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.multiHMM}
\alias{plot.multiHMM}
\title{Plotting function for \code{\link{multiHMM}} objects}
\usage{
\method{plot}{multiHMM}(x, type = "transitionMatrix", ...)
}
\arguments{
\item{x}{A \code{\link{multiHMM}} object.}

\item{type}{Type of the plot, one of \code{c('transitionMatrix','histograms','correlation')}. You can also specify the type with an integer number.
\describe{
  \item{\code{transitionMatrix}}{A heatmap with entries of the transition matrix.}
  \item{\code{histograms}}{Fitted histograms of all underlying univariate distributions.}
  \item{\code{correlation}}{Correlation between read counts of the different experiments.}
}}

\item{...}{Additional arguments for the different plot types.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make different types of plots for \code{\link{multiHMM}} objects.
}

