% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCombinatorialEntropy}
\alias{plotCombinatorialEntropy}
\title{Plot combinatorial entropy}
\usage{
plotCombinatorialEntropy(gr, genome)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object generated by \code{\link{combinatorialEntropy}}}

\item{genome}{The genome assembly to use. Type \code{rtracklayer::ucscGenomes()$db} for a list of available genomes.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a karyogram with combinatorial entropy generated by \code{\link{combinatorialEntropy}}.
}
\author{
Aaron Taudt
}

