% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot enrichment around annotation}
\usage{
plotEnrichment(hmm, annotation, bp.around.annotation = 10000,
  region = c("start", "inside", "end"), num.intervals = 20)
}
\arguments{
\item{hmm}{A \code{\link{multiHMM}} object.}

\item{annotation}{A \code{\link{GRanges}} object with the annotation of interest.}

\item{bp.around.annotation}{An integer specifying the number of basepairs up- and downstream of the annotation for which the enrichment will be calculated.}

\item{region}{A combination of \code{c('start','inside','end')} specifying the region of the annotation for which the enrichment will be calculated. Select \code{'start'} if you have a point-sized annotation like transcription start sites. Select \code{c('start','inside','end')} if you have long annotations like genes.}

\item{num.intervals}{Number of intervals for enrichment 'inside' of annotation.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object containing the plot.
}
\description{
Plot fold enrichment of combinatorial states around and inside of annotation.
}
\author{
Aaron Taudt
}

