% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionAnalysis.R
\name{plotExpression}
\alias{plotExpression}
\title{Overlap with expression data}
\usage{
plotExpression(hmm, expression, combinations = NULL, return.marks = FALSE)
}
\arguments{
\item{hmm}{A \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object or file that contains such an object.}

\item{expression}{A \code{\link{GRanges}} object with metadata column 'expression', containing the expression value for each range.}

\item{combinations}{A vector with combinations for which the expression overlap will be calculated. If \code{NULL} all combinations will be considered.}

\item{return.marks}{Set to \code{TRUE} if expression values for marks instead of combinations should be returned.}
}
\value{
A named list with expression values.
}
\description{
Get the expression values that overlap with each combinatorial state.
}
\examples{
## Load an example multiHMM
file <- system.file("data","multivariate_mode-combinatorial_condition-SHR.RData",
                    package="chromstaR")
model <- get(load(file))

## Obtain expression data
data(expression_lv)
head(expression_lv)

## We need to get coordinates for each of the genes
library(biomaRt)
ensembl <- useMart('ENSEMBL_MART_ENSEMBL', host='may2012.archive.ensembl.org',
                  dataset='rnorvegicus_gene_ensembl')
genes <- getBM(attributes=c('ensembl_gene_id', 'chromosome_name', 'start_position',
                           'end_position', 'strand', 'external_gene_id',
                           'gene_biotype'),
              mart=ensembl)
expr <- merge(genes, expression_lv, by='ensembl_gene_id')
# Transform to GRanges
expression.SHR <- GRanges(seqnames=paste0('chr',expr$chromosome_name),
                         ranges=IRanges(start=expr$start, end=expr$end),
                         strand=expr$strand, name=expr$external_gene_id,
                         biotype=expr$gene_biotype,
                         expression=expr$expression_SHR)
# We apply an asinh transformation to reduce the effect of outliers
expression.SHR$expression <- asinh(expression.SHR$expression)

## Plot
plotExpression(model, expression.SHR) +
 theme(axis.text.x=element_text(angle=0, hjust=0.5)) +
 ggtitle('Expression of genes overlapping combinatorial states')
plotExpression(model, expression.SHR, return.marks=TRUE) +
 ggtitle('Expression of marks overlapping combinatorial states')

}
\author{
Aaron Taudt
}
\seealso{
\code{\link{plotting}}
}

