% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot read counts around annotation}
\usage{
plotHeatmap(hmm, annotation, bp.around.annotation = 10000, max.rows = 1000)
}
\arguments{
\item{hmm}{A \code{\link{multiHMM}} object.}

\item{annotation}{A \code{\link{GRanges}} object with the annotation of interest.}

\item{bp.around.annotation}{An integer specifying the number of basepairs up- and downstream of the annotation for which the enrichment will be calculated.}

\item{max.rows}{An integer specifying the number of randomly subsampled rows that are plotted from the \code{annotation} object. This is necessary to avoid crashing for heatmaps with too many rows.}
}
\description{
Plot read counts around annotation as heatmap.
}
\author{
Aaron Taudt
}

