% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHistograms}
\alias{plotHistograms}
\title{Histograms of binned read counts with fitted mixture distribution}
\usage{
plotHistograms(multi.hmm, ...)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{state}{Plot the histogram only for the specified state. One of \code{c('unmodified','modified')}.}

\item{chromosomes, start, end}{Plot the histogram only for the specified chromosomes, start and end position.}

\item{linewidth}{Width of the distribution lines.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot histograms of binned read counts with fitted mixture distributions from a \code{\link{multiHMM}} object.
}

