% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotKaryogram}
\alias{plotKaryogram}
\title{Plot a karyogram with read counts and univariate peak calls}
\usage{
plotKaryogram(model)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} object or file that contains such an object.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a karyogram with read counts and peak calls from a \code{\link{uniHMM}} object.
}
\author{
Aaron Taudt
}
