\name{plotting functions for chromstar}
\alias{plot.distribution}
\alias{plot.boxplot}
\alias{plot.distribution.normal}

\title{
Plotting functions for the \code{\link{chromstar}} package
}
\description{
Plot histogram of read counts or boxplots for univariate fits. Plot transformed read counts for univariate fits.
}
\usage{
plot.distribution(model, state=NULL, chr=NULL, start=NULL, end=NULL)
}
\arguments{
\item{model}{
A univariate HMM of class \code{\link{chromstar.univariate.hmm}} generated by \code{\link{univariate.from.binned.data}}.
}
\item{state}{
Either "unmodified" or "modified". Plots only the histogram for bins that belong to the specified state.
}
\item{chr}{
A string describing a chromosome, e.g. "chr1". This will plot only the histogram and fits for this chromosome.
}
\item{start}{
An integer specifying the start position of the region to plot. If chromosome is unset and there is more than one chromosome in the model, this option will be ignored.
}
\item{end}{
An integer specifying the end position of the region to plot. If chromosome is unset and there is more than one chromosome in the model, this option will be ignored.
}
}

\details{
}
\keyword{histogram}
\keyword{distributions}
\keyword{univariate}
\references{}
\author{Aaron Taudt}
\note{}
\seealso{}
\examples{
# Load example datasets with genome coordinates and number of aligned reads for each bin
data(binned.data.example.1)
# Do the univariate HMMs
model = univariate.from.binned.data(binned.data.example.1, eps=1, max.time=60)
# Plot histogram of binned reads and fitted distributions
plot.distribution(model)
}

