% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTransitionProbs}
\alias{plotTransitionProbs}
\title{Heatmap of transition probabilities}
\usage{
plotTransitionProbs(model)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a heatmap of transition probabilities for a \code{\link{multiHMM}} model.
}

