% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotUnivariateHistogram}
\alias{plotUnivariateHistogram}
\title{Plot a histogram of binned read counts with fitted mixture distribution}
\usage{
plotUnivariateHistogram(model, state = NULL, chromosomes = NULL,
  start = NULL, end = NULL, linewidth = 1)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} object.}

\item{state}{Plot the histogram only for the specified state. One of \code{c('unmodified','modified')}.}

\item{chromosomes, start, end}{Plot the histogram only for the specified chromosomes, start and end position.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a histogram of binned read counts from with fitted mixture distributions from a \code{\link{uniHMM}} object.
}

