% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotUnivariateKaryogram}
\alias{plotUnivariateKaryogram}
\title{Plot a karyogram with read counts and univariate peak calls}
\usage{
plotUnivariateKaryogram(model)
}
\arguments{
\item{model}{A \code{\link{chromstaR_univariateHMM}} object or file that contains such an object.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a karyogram with read counts and peak calls from a \code{\link{chromstaR_univariateHMM}} object.
}
\author{
Aaron Taudt
}

