% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.combinedMultiHMM}
\alias{print.combinedMultiHMM}
\title{Print combinedMultiHMM object}
\usage{
\method{print}{combinedMultiHMM}(x, ...)
}
\arguments{
\item{x}{An \code{\link{combinedMultiHMM}} object.}

\item{...}{Ignored.}
}
\value{
An invisible \code{NULL}.
}
\description{
Print combinedMultiHMM object
}
