% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.multiHMM}
\alias{print.multiHMM}
\title{Print multiHMM object}
\usage{
\method{print}{multiHMM}(x, ...)
}
\arguments{
\item{x}{An \code{\link{multiHMM}} object.}

\item{...}{Ignored.}
}
\value{
An invisible \code{NULL}.
}
\description{
Print multiHMM object
}

