% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.uniHMM}
\alias{print.uniHMM}
\title{Print uniHMM object}
\usage{
\method{print}{uniHMM}(x, ...)
}
\arguments{
\item{x}{An \code{\link{uniHMM}} object.}

\item{...}{Ignored.}
}
\value{
An invisible \code{NULL}.
}
\description{
Print uniHMM object
}
