% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedimport.R
\name{readBedFile}
\alias{readBedFile}
\title{Read bed-file into GRanges}
\usage{
readBedFile(bedfile, skip = 0)
}
\arguments{
\item{bedfile}{Filename of the bed or bed.gz file.}

\item{skip}{Number of lines to skip at the beginning.}
}
\value{
A \code{\link{GRanges}} object with the contents of the bed-file.
}
\description{
This is a simple convenience function to read a bed(.gz)-file into a \code{\link{GRanges}} object. The bed-file is expected to have at least the following three fields: \code{chromosome, start, end}.
}
\examples{
## Get an example BED file
bedfile <- system.file("extdata", "liver-H3K4me3-BN-male-bio2-tech1.bed.gz",
                       package="chromstaRData")
## Import the file and skip the first 10 lines
data <- readBedFile(bedfile, skip=10)

}
\author{
Aaron Taudt, David Widmann
}

