% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importReads.R
\name{readBedFileAsGRanges}
\alias{readBedFileAsGRanges}
\title{Import BED file into GRanges}
\usage{
readBedFileAsGRanges(bedfile, assembly, chromosomes = NULL,
  remove.duplicate.reads = FALSE, min.mapq = 10,
  max.fragment.width = 1000, blacklist = NULL)
}
\arguments{
\item{bedfile}{A file with aligned reads in BED-6 format. The columns have to be c('chromosome','start','end','description','mapq','strand').}

\item{assembly}{Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Only necessary when importing BED files. BAM files are handled automatically. Alternatively a data.frame with columns 'chromosome' and 'length'.}

\item{chromosomes}{If only a subset of the chromosomes should be imported, specify them here.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=0} to keep all reads.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments.}

\item{blacklist}{A \code{\link{GRanges}} or a bed(.gz) file with blacklisted regions. Reads falling into those regions will be discarded.}
}
\value{
A \code{\link{GRanges}} object containing the reads.
}
\description{
Import aligned reads from a BED file into a \code{\link{GRanges}} object.
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata", "liver-H3K4me3-BN-male-bio2-tech1.bed.gz",
                       package="chromstaRData")
## Read the file into a GRanges object
data(rn4_chrominfo)
reads <- readBedFileAsGRanges(bedfile, assembly=rn4_chrominfo, chromosomes='chr12',
                    min.mapq=10, remove.duplicate.reads=TRUE)
print(reads)

}
