% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCondition.R
\name{removeCondition}
\alias{removeCondition}
\title{Remove condition from model}
\usage{
removeCondition(model, conditions)
}
\arguments{
\item{model}{A \code{\link{combinedMultiHMM}} object or file which contains such an object.}

\item{conditions}{A character vector with the condition(s) to be removed.}
}
\value{
The input \code{\link{combinedMultiHMM}} object with specified conditions removed.
}
\description{
Remove a condition from a \code{\link{combinedMultiHMM}} object.
}
\examples{
## Get an example HMM
file <- system.file("data","combined_mode-differential.RData",
                    package="chromstaR")
model <- get(load(file))

## Print available conditions
print(unique(model$info$condition))

## Remove condition SHR
new.model <- removeCondition(model, conditions='SHR')

}
