% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\alias{differentialScoreMax}
\alias{differentialScoreSum}
\title{chromstaR scores}
\usage{
differentialScoreMax(mat, info, FUN = "-")

differentialScoreSum(mat, info, FUN = "-")
}
\arguments{
\item{mat}{A matrix with posterior probabilities, read counts or any other matrix with these dimensions. Column names must correspond to the ID entries in \code{info}.}

\item{info}{An \code{\link{experiment.table}} with additional column 'ID'.}

\item{FUN}{A function to compute the score with.}
}
\value{
A numeric vector.
}
\description{
Various scores used in \code{\link{chromstaR}}.
}
\section{Functions}{
\itemize{
\item \code{differentialScoreMax}: Maximum differential score. Values are between 0 and 1. A value of 1 means that at least one mark is maximally different between conditions.

\item \code{differentialScoreSum}: Additive differential score. Values are between 0 and N, where N is the number of marks. A value around 1 means that approximately 1 mark is different, a value of 2 means that 2 marks are different etc.
}}

\author{
Aaron Taudt
}
