% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateReadsFromCounts}
\alias{simulateReadsFromCounts}
\title{Simulate read coordinates}
\usage{
simulateReadsFromCounts(bins, fragLen = 50)
}
\arguments{
\item{bins}{A \code{\link[GenomicRanges]{GRanges}} with read counts.}

\item{fragLen}{Length of the simulated read fragments.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} with read coordinates.
}
\description{
Simulate read coordinates using read counts as input.
}
