\name{summarize}
\alias{summarize}
\title{
Summary information
}
\description{
Get some information about a \code{\link{chromstaR_multivariateHMM}} object.
}
\usage{
summarize(multi.hmm)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} object.}
}

\details{
Gives summary information about differential states and peaks per sample.
}
\value{
A list object with various entries.
}
\keyword{summary}
\author{Aaron Taudt}
\examples{
}
