% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyAnalysis.R
\name{transitionFrequencies}
\alias{transitionFrequencies}
\title{Transition frequencies of combinatorial states}
\usage{
transitionFrequencies(
  multi.hmms = NULL,
  combined.hmm = NULL,
  zero.states = "[]",
  combstates = NULL
)
}
\arguments{
\item{multi.hmms}{A named list with \code{\link{multiHMM}} objects or a vector with filenames that contain such objects.}

\item{combined.hmm}{A \code{\link{combinedMultiHMM}} object. If specified, \code{multi.hmms} is ignored.}

\item{zero.states}{The string(s) which identifies the zero.states.}

\item{combstates}{Alternative input instead of \code{multi.hmms}: A named list of combinatorial state vectors instead of HMMs. Names must be of the form "combination.X", where X is an arbitrary string. If this is specified, \code{multi.hmms} and \code{combined.hmm} will be ignored.}
}
\value{
A data.frame with transition frequencies.
}
\description{
Get a table of transition frequencies between combinatorial states of different \code{\link{multiHMM}}s.
}
\examples{
## Get an example combinedMultiHMM
file <- system.file("data","combined_mode-differential.RData",
                    package="chromstaR")
model <- get(load(file))
freqs <- transitionFrequencies(combined.hmm=model)
freqs$table

}
\author{
Aaron Taudt
}
