% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScore.R
\name{addScore}
\alias{addScore}
\title{Add score for differential export}
\usage{
addScore(multi.model, replicates, conditions, tracks2compare,
  reorderSegments = TRUE)
}
\description{
Add column 'score' to the 'segments' entry in a \code{\link{multiHMM}}. This score can be used to rank differential states.
}
\author{
Aaron Taudt
}

