% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2GRanges.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{Read a BED file into a \link{GRanges} objects.}
\usage{
bed2GRanges(bedfile, chrom.length.file, skip = 1, binsize = NULL)
}
\arguments{
\item{bedfile}{A BED file.}

\item{chrom.length.file}{A file which contains the chromosome lengths in basepairs. The first column contains the chromosome name and the second column the length (see also \code{\link{chrom.length.file}}.}

\item{skip}{Number of lines to skip in BED file.}

\item{binsize}{Force the resulting \link{GRanges} to have this bin size.}
}
\value{
A \link{GRanges} object.
}
\description{
Read a BED file into a \link{GRanges} object. Only coordinates are read.
}
\author{
Aaron Taudt
}

