% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{binned.data}
\alias{binned.data}
\title{Binned read counts}
\description{
A \code{\link[GenomicRanges]{GRanges}} object which contains binned read counts as meta data column \code{counts}. It is output of the \code{\link{binReads}} function.
}
