% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{bam2binned}
\alias{bed2binned}
\alias{bedGraph2binned}
\alias{binning}
\title{Convert aligned reads from various file formats into read counts in equidistant bins}
\usage{
bam2binned(bamfile, bamindex = bamfile, pairedEndReads = FALSE,
  outputfolder = "binned_data", binsizes = 500, chromosomes = NULL,
  save.as.RData = FALSE, min.mapq = 10, downsample.to.reads = NULL,
  remove.duplicate.reads = FALSE, max.fragment.width = 1000)

bed2binned(bedfile, assembly, chrom.length.file = NULL,
  outputfolder = "binned_data", binsizes = 500, chromosomes = NULL,
  save.as.RData = FALSE, downsample.to.reads = NULL)

bedGraph2binned(bedGraphfile, assembly, chrom.length.file = NULL,
  outputfolder = "binned_data", binsizes = 500, chromosomes = NULL,
  save.as.RData = FALSE, downsample.to.reads = NULL)
}
\arguments{
\item{bamfile}{A file in BAM format.}

\item{bamindex}{BAM index file. Can be specified without the .bai ending. If this file does not exist it will be created and a warning is issued.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{outputfolder}{Folder to which the binned data will be saved if \code{save.as.RData=TRUE}. If the specified folder does not exist, it will be created.}

\item{binsizes}{A vector with integer values which will be used for the binning. If more than one value is given, output files will be produced for each bin size.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{save.as.RData}{If set to \code{TRUE}, output will be written to file. The filename is generated from the input file and bin sizes. If set to \code{FALSE}, no output file will be written. Instead, a \link{GenomicRanges} object containing the binned data will be returned. Only the first binsize will be processed in this case.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{downsample.to.reads}{Downsample the input data to the specified number of reads.}

\item{remove.duplicate.reads}{Logical indicating whether duplicate reads should be removed. NOTE: Duplicate removal only works if your duplicate reads are properly flagged in the BAM file.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments due to mapping errors of paired end reads.}

\item{bedfile}{A file in BED format.}

\item{assembly}{An assembly to specify the chromosome lengths. One of \code{c('hg19','hg18')}.}

\item{chrom.length.file}{A file which contains the chromosome lengths in basepairs. The first column contains the chromosome name and the second column the length (see also \code{\link{chrom.length.file}}.}

\item{bedGraphfile}{A file in bedGraph format.}
}
\description{
Convert aligned reads in .bam or .bed format into read counts in equidistant windows. Convert signal values in .bedGraph format to signal counts in equidistant windows.
}
\details{
Convert aligned reads or signal values from various file formats into read counts in equidistant windows (bins). bam2binned() and bed2binned() use 'GenomicRanges::countOverlaps()' to calculate the read counts. Therefore, with small binsizes and large read lengths, one read fragment will often overlap more than one bin.
bedGraph2binned() sets the maximum signal value in a bin as value for that bin.
}
\section{Functions}{
\itemize{
\item \code{bam2binned}: Bin reads in BAM format.

\item \code{bed2binned}: Bin reads in BED format.

\item \code{bedGraph2binned}: Bin reads in bedGraph format.
}}
\examples{
## Get an example BED-file with ChIP-seq reads for H3K36me3 in brain tissue
bedfile <- system.file(
     "extdata/brain/BI.Brain_Angular_Gyrus.H3K36me3.112.chr22.bed.gz",
     package="chromstaR")
## Bin the BED files into bin size 1000bp
binned.data <- bed2binned(bedfile, assembly='hg19', binsize=1000,
                         save.as.RData=FALSE)
## Get the number of bins
length(binned.data)
}
\author{
Aaron Taudt
}
\seealso{
\link{binned.data}
}

