\name{callPeaksMultivariate}
\alias{callPeaksMultivariate}
\title{
Fit a Hidden Markov Model to multiple ChIP-seq samples.
}
\description{
Fit a HMM to multiple ChIP-seq samples to determine the combinatorial state of genomic regions. Input is a list of \code{\link{chromstaR_univariateHMM}}s generated by \code{\link{callPeaksUnivariate}}.
}
\usage{
callPeaksMultivariate(modellist, use.states=NULL, num.states=NULL, eps=0.001, FDR=NULL, keep.posteriors=FALSE, num.threads=1, max.time=-1, max.iter=-1, checkpoint.after.iter=-1, checkpoint.after.time=-1, checkpoint.file=NULL, checkpoint.overwrite=TRUE, checkpoint.use.existing=FALSE, A.initial=NULL)
}
\arguments{
\item{modellist}{
A list of \code{\link{chromstaR_univariateHMM}}s generated by \code{\link{callPeaksUnivariate}}, e.g. \code{list(hmm1,hmm2,...)} or \code{c("file1","file2",...)}.
}
\item{use.states}{
A vector of combinatorial states which are used in the multivariate HMM. If specified, option \code{num.states} will be ignored. An error is thrown if the specified states cannot be used (either because they do not exist or their occurrence in the genome is too low).
} 
\item{num.states}{
Number of combinatorial states to use in the multivariate HMM. The states are ordered by occurrence as determined from the combination of univariate state calls. If it is not possible to use the specified number of states - e.g. because the total number of combinatorial states is lower or because the inverse of the correlation matrix would be singular for some states - a warning is issued and the maximum possible number of states is used instead.
}
\item{eps}{Convergence threshold for the Baum-Welch algorithm.}
\item{FDR}{False discovery rate. NULL means that the state with maximum posterior probability will be chosen, irrespective of its absolute probability (default=NULL).}
\item{keep.posteriors}{If set to TRUE (default=FALSE), posteriors will be available in the output. This is useful to change the FDR later, but increases the necessary disk space to store the result immense.}

\item{num.threads}{Number of threads to use. Setting this to >1 may give increased performance.}
\item{max.time}{The maximum running time in seconds for the Baum-Welch algorithm. If this time is reached, the Baum-Welch will terminate after the current iteration finishes. The default -1 is no limit.}
\item{max.iter}{The maximum number of iterations for the Baum-Welch algorithm. The default -1 is no limit.}
\item{checkpoint.after.iter}{Write a checkpoint file every n iterations. The default -1 means no checkpointing for iterations.}
\item{checkpoint.after.time}{Write a checkpoint file every t seconds. The default -1 means no checkpointing for time.}
\item{checkpoint.file}{The name of the checkpoint file that will be written.}
\item{checkpoint.overwrite}{
If set to TRUE, only one checkpoint file will be written. If set to FALSE, a new checkpoint file will be written at each checkpoint with the total number of iterations appended.
}
\item{checkpoint.use.existing}{If set to TRUE, the Baum-Welch fitting procedure will be continued from the HMM in the \code{checkpoint.file}.}
\item{A.initial}{A transition matrix to start the Baum-Welch from.}
}

\details{
Emission distributions from the univariate HMMs are used with a Gaussian copula to generate a multivariate emission distribution for each combinatorial state. This multivariate distribution is then kept fixed and the transition probabilities are fitted with a Baum-Welch. See our paper TODO:insert paper for details.
}
\value{
Output is a \code{\link{chromstaR_multivariateHMM}} list object.
}

\keyword{Hidden Markov Model}
\keyword{Baum-Welch}
\keyword{multivariate}
\references{TODO:insert paper}
\author{Aaron Taudt and Maria Colome Tatche}
\seealso{\code{\link{chromstaR_multivariateHMM}}, \code{\link{callPeaksUnivariate}}}
\examples{
}
