% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFDR.R
\name{changeFDR}
\alias{changeFDR}
\title{Change the false discovery rate of a Hidden Markov Model}
\usage{
changeFDR(model, fdr = 0.01, invert = FALSE)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} or \code{\link{multiHMM}} object with posteriors.}

\item{fdr}{A vector of values between 0 and 1 for each column in \code{model$bins$peakScores}. If only one value is given, it will be reused for all columns. Values close to 0 will yield more stringent peak calls with lower false positive but higher false negative rate.}

\item{invert}{Select peaks below (\code{FDR}) or above (\code{TRUE}) the given \code{fdr}. This is useful to select low confidence peaks.}
}
\value{
The input object is returned with adjusted peak calls.
}
\description{
Adjusts the peak calls of a \code{\link{uniHMM}} or \code{\link{multiHMM}} object with the given false discovery rate.
}
\details{
Each peak has a peak score associated (between 0 and 1). The false discovery rate is a simple cutoff on 1-peakScore, e.g. for an \code{fdr = 0.01} only peaks with a peak score \code{>= 0.99} (1-fdr) will be selected.
}
\examples{
## Get an example BAM file
file <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bam",
                       package="chromstaRData")
## Bin the file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(file, assembly=rn4_chrominfo, binsizes=1000,
                  chromosomes='chr12')
## Fit the univariate Hidden Markov Model
hmm <- callPeaksUnivariate(binned, max.time=60, eps=1)
## Compare fits with different fdrs
plotHistogram(hmm) + ylim(0,0.25) + ylim(0,0.3)
plotHistogram(changeFDR(hmm, fdr=0.01)) + ylim(0,0.3)
plotHistogram(changeFDR(hmm, fdr=1e-12)) + ylim(0,0.3)

}
\author{
Aaron Taudt
}

