% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changePostCutoff.R
\name{changeMaxPostCutoff}
\alias{changeMaxPostCutoff}
\alias{changeFDR}
\title{Adjust sensitivity of peak detection}
\usage{
changeMaxPostCutoff(model, maxPost.cutoff = 0.99, invert = FALSE)

changeFDR(model, fdr = 0.01, invert = FALSE)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} or \code{\link{multiHMM}} object with posteriors.}

\item{maxPost.cutoff}{A vector of values between 0 and 1 for each column in \code{model$bins$posteriors}. If only one value is given, it will be reused for all columns. Values close to 1 will yield more stringent peak calls with lower false positive but higher false negative rate (i.e. more precise but less sensitive).}

\item{invert}{Select peaks below (\code{FALSE}) or above (\code{TRUE}) the given \code{maxPost.cutoff}. This is useful to select low confidence peaks.}

\item{fdr}{Same as \code{1-maxPost.cutoff}.}
}
\value{
The input object is returned with adjusted peak calls.
}
\description{
Adjusts the peak calls of a \code{\link{uniHMM}}, \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object with a cutoff on the maximum-posterior within each peak. Higher values of \code{maxPost.cutoff} mean less sensitive and more precise peak calls. Remaining peaks are kept intact, as opposed to function \code{\link{changePostCutoff}}, where broad peaks are fragmented. This function was formerly called 'changeFDR' and is still available for backwards compatibiltiy.
}
\details{
Each peak has a maximum-posterior (maxPostInPeak, between 0 and 1) associated. The sensitivity is adjusted with a simple cutoff on maxPostInPeak, e.g. for \code{maxPost.cutoff = 0.99} only peaks with \code{maxPostInPeak >= 0.99} will be selected.
}
\section{Functions}{
\itemize{
\item \code{changeFDR}: This function was renamed to 'changeMaxPostCutoff' in chromstaR 1.5.1 but it still available for backwards compatibility.
}}

\examples{
## Get an example BAM file
file <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bam",
                       package="chromstaRData")
## Bin the file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(file, assembly=rn4_chrominfo, binsizes=1000,
                  stepsizes=500, chromosomes='chr12')
## Fit the univariate Hidden Markov Model
hmm <- callPeaksUnivariate(binned, max.time=60, eps=1)
## Compare fits with different fdrs
plotHistogram(hmm) + ylim(0,0.25) + ylim(0,0.3)
plotHistogram(changeMaxPostCutoff(hmm, maxPost.cutoff=0.99)) + ylim(0,0.3)
plotHistogram(changeMaxPostCutoff(hmm, maxPost.cutoff=1-1e-12)) + ylim(0,0.3)

}
\seealso{
\code{\link{changePostCutoff}}
}
\author{
Aaron Taudt
}
