% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changePostCutoff.R
\name{changePostCutoff}
\alias{changePostCutoff}
\title{Change the false discovery rate of a Hidden Markov Model}
\usage{
changePostCutoff(model, post.cutoff = 0.5, separate.zeroinflation = TRUE,
  averages = TRUE)
}
\arguments{
\item{model}{A \code{\link{uniHMM}} or \code{\link{multiHMM}} object with posteriors.}

\item{post.cutoff}{Posterior cutoff. Values close to 1 will yield more stringent peak calls with lower false positive but higher false negative rate.}

\item{separate.zeroinflation}{Only for \code{\link{uniHMM}} objects: If set to TRUE, state 'zero-inflation' will be treated separately, otherwise it will be merged with state 'unmodified'.}

\item{averages}{Whether or not averaged posteriors should appear in the segmentation.}
}
\value{
The input object is returned with adjusted peak calls.
}
\description{
Adjusts the peak calls of a \code{\link{uniHMM}} or \code{\link{multiHMM}} object with the given posterior cutoff.
}
\details{
Posterior probabilities are between 0 and 1. Peaks are called if the posteriors for a state (univariate) or sample (multivariate) are >= \code{post.cutoff}.
}
\examples{
## Get an example BED file
bedfile <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bed.gz",
                       package="chromstaRData")
## Bin the BED file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(bedfile, assembly=rn4_chrominfo, binsize=1000,
                  chromosomes='chr12')
## Fit the univariate Hidden Markov Model
# !Keep posteriors to change the post.cutoff later!
hmm <- callPeaksUnivariate(binned, ID='example_H3K27me3', max.time=60, eps=1,
                          keep.posteriors=TRUE)
## Compare fits with different post.cutoffs
plot(changePostCutoff(hmm, post.cutoff=0.01), type='histogram') + ylim(0,0.25)
plot(hmm, type='histogram') + ylim(0,0.25)
plot(changePostCutoff(hmm, post.cutoff=0.99), type='histogram') + ylim(0,0.25)

}
\author{
Aaron Taudt
}

