% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromstaR-package.R
\docType{package}
\name{chromstaR-package}
\alias{chromstaR}
\alias{chromstaR-package}
\title{Combinatorial and differential ChIP-seq analysis}
\description{
This package implements functions for the combinatorial and differential analysis of ChIP-seq data. It was developed for histone modifications with a broad profile but is also suitable for the analysis of transcription factor binding data. A Hidden Markov Model with a mixture of Negative Binomials as emission densities is used to call peaks. Please read our publication for a detailed description of the method. TODO: insert publication.
}
\details{
Please read \href{../doc/chromstaR.pdf}{The chromstaR user's guide} for a quick introduction.
}
\author{
Aaron Taudt, Maria Colome-Tatche, Matthias Heinig, Minh Anh Nguyen
}

