% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromstaR-package.R
\docType{package}
\name{chromstaR-package}
\alias{chromstaR}
\alias{chromstaR-package}
\title{Combinatorial and differential ChIP-seq analysis}
\description{
This package implements functions for the combinatorial and differential analysis of ChIP-seq data.
}
\details{
Please read \href{../doc/chromstaR.pdf}{The chromstaR user's guide} for a quick introduction.
}
\author{
Aaron Taudt, Maria Colome-Tatche, Matthias Heinig, Minh Anh Nguyen
}

