% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combinatorialEntropy.R
\name{combinatorialEntropy}
\alias{combinatorialEntropy}
\title{Combinatorial entropy}
\usage{
combinatorialEntropy(multi.hmm.list, direction.of.adding = "between.hmms")
}
\arguments{
\item{multi.hmm.list}{A list of \code{\link{chromstaR_multivariateHMM}} objects or a vector of files that contain such objects.}

\item{direction.of.adding}{One of \code{c('between.hmms','between.samples')}.}
}
\description{
Calculate the combinatorial entropy along the genome for a set of \code{\link{chromstaR_multivariateHMM}}s.
}
\details{
The combinatorial entropy is calculated as ...
}
\author{
Aaron Taudt
}

