% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{combinatorialFrequency}
\alias{combinatorialFrequency}
\title{Frequency of combinatorial states}
\usage{
combinatorialFrequency(multi.hmm, combinations = NULL)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} or a file that contains such an object.}

\item{combinations}{A vector with combinations for which the frequency will be calculated. If \code{NULL} all combinations will be considered.}
}
\value{
A table with frequencies of each combinatorial state.
}
\description{
Get the genomewide frequency of each combinatorial state.
}
\author{
Aaron Taudt
}

