% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMultivariates.R
\name{combineMultivariates}
\alias{combineMultivariates}
\title{Combine combinatorial states from several Multivariates}
\usage{
combineMultivariates(multi.hmm.list, mode, conditions)
}
\arguments{
\item{multi.hmm.list}{A named \code{list()} with \code{\link{multiHMM}} objects. The names of the list are used to name the conditions or marks. Alternatively a named character vector with filenames that contain \code{\link{multiHMM}} objects.}

\item{mode}{Mode of combination. See \code{\link{Chromstar}} for a description of the \code{mode} parameter.}

\item{conditions}{A character vector with conditions in case of \code{mode='full'} or \code{mode='condition'}.}
}
\value{
A \code{link{combinedMultiHMM}} objects with combinatorial states for each condition.
}
\description{
Combine combinatorial states from several \code{\link{multiHMM}} objects. Combinatorial states can be combined for objects containing multiple marks (\code{mode='mark'}) or multiple conditions (\code{mode='condition'}).
}
\examples{
### Multivariate peak calling for spontaneous hypertensive rat (SHR) ###
# Get example BED files for 2 different marks in hypertensive rat (SHR)
file.path <- system.file("extdata","euratrans", package='chromstaRData')
bedfiles <- list.files(file.path, full.names=TRUE, pattern='SHR')[c(1:2,6:7)]
# Bin the data
data(rn4_chrominfo)
binned.data <- list()
for (bedfile in bedfiles) {
 binned.data[[basename(bedfile)]] <- binReads(bedfile, binsize=1000,
                                              assembly=rn4_chrominfo, chromosomes='chr12')
}
# Obtain the univariate fits
models <- list()
for (i1 in 1:length(binned.data)) {
 models[[i1]] <- callPeaksUnivariate(binned.data[[i1]], ID=names(binned.data)[i1],
                                     max.time=60, eps=1)
}
# Construct experiment structure
exp <- data.frame(file=bedfiles, mark=c("H3K27me3","H3K27me3","H3K4me3","H3K4me3"),
                 condition=rep("SHR",4), replicate=c(1:2,1:2), pairedEndReads=FALSE)
states <- stateBrewer(exp, mode='mark')
# Call multivariate peaks
multimodel.SHR <- callPeaksMultivariate(models, use.states=states, eps=1, max.time=60)

#'### Multivariate peak calling for brown norway (BN) rat ###
# Get example BED files for 2 different marks in brown norway rat
file.path <- system.file("extdata","euratrans", package='chromstaRData')
bedfiles <- list.files(file.path, full.names=TRUE, pattern='BN')[c(1:2,7:8)]
# Bin the data
data(rn4_chrominfo)
binned.data <- list()
for (bedfile in bedfiles) {
 binned.data[[basename(bedfile)]] <- binReads(bedfile, binsize=1000,
                                              assembly=rn4_chrominfo, chromosomes='chr12')
}
# Obtain the univariate fits
models <- list()
for (i1 in 1:length(binned.data)) {
 models[[i1]] <- callPeaksUnivariate(binned.data[[i1]], ID=names(binned.data)[i1],
                                     max.time=60, eps=1)
}
# Construct experiment structure
exp <- data.frame(file=bedfiles, mark=c("H3K27me3","H3K27me3","H3K4me3","H3K4me3"),
                 condition=rep("BN",4), replicate=c(1:2,1:2), pairedEndReads=FALSE)
states <- stateBrewer(exp, mode='mark')
# Call multivariate peaks
multimodel.BN <- callPeaksMultivariate(models, use.states=states, eps=1, max.time=60)

### Combine multivariates ###
comb.model <- combineMultivariates(list(SHR=multimodel.SHR, BN=multimodel.BN), mode='mark')

}
\author{
Aaron Taudt
}

