% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMultivariates.R
\name{combineMultivariates}
\alias{combineMultivariates}
\title{Combine combinatorial states from several Multivariates}
\usage{
combineMultivariates(conditions = NULL, marks = NULL)
}
\arguments{
\item{conditions}{A named \code{list()} with \code{\link{multiHMM}} objects. The names of the list are used to name the conditions. Alternatively a named character vector with filenames that contain \code{\link{multiHMM}} objects.}

\item{marks}{A named \code{list()} with \code{\link{multiHMM}} objects. The names of the list are used to name the marks. Alternatively a named character vector with filenames that contain \code{\link{multiHMM}} objects.}
}
\value{
A DataFrame with combinatorial states for each condition.
}
\description{
Combine combinatorial states from several \code{\link{multiHMM}} objects. Combinatorial states can be combined across \code{conditions} (from a combinatorial analysis) or across \code{marks} (from a differential analysis).
}
\author{
Aaron Taudt
}

