% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{example.multi.HMM}
\alias{example.multi.HMM}
\title{Multivariate Hidden Markov Model for usage in examples}
\format{See \code{\link{chromstaR_multivariateHMM}} for a description of the format.}
\description{
A \code{\link{chromstaR_multivariateHMM}} object, generated from the brain tissue dataset (H3K36me3, 7 tissues) in 'extdata/brain/'. This dataset is used in the example section of \pkg{chromstaR} functions.
}

