% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.R
\name{exportBinnedData}
\alias{exportBinnedData}
\title{Export genome browser viewable files}
\usage{
exportBinnedData(binned.data.list, filename = "chromstaR_ReadCounts",
  header = TRUE, separate.files = FALSE)
}
\arguments{
\item{binned.data.list}{A list of \code{\link{binned.data}} objects or files that contain such objects.}

\item{filename}{The name of the file that will be written. The ending ".wig.gz" for read counts will be appended. Any existing file will be overwritten.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Export read counts as genome browser viewable file
}
\details{
Export read counts from \code{\link{binned.data}} as a file which can be uploaded into a genome browser. Read counts are exported in WIGGLE format (.wig.gz).
}
\examples{
## Get example BED-files with ChIP-seq reads for H3K36me3
# in 7 different brain tissues (chr22)
bedfiles <- list.files(system.file(file.path("extdata","brain"),
                      package="chromstaR"), full=TRUE)
## Bin the data into bin size 1000bp
binned.data.list <- list()
for (bedfile in bedfiles) {
 binned.data.list[[bedfile]] <- bed2binned(bedfile, assembly='hg19', binsize=1000,
                                           save.as.RData=FALSE)
}
## Export the binned read counts
\donttest{exportBinnedData(binned.data.list, filename='chromstaR-example_binned.data')}
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportMultivariate}}
}

