% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportCombinedMultivariate}
\alias{exportCombinedMultivariate}
\title{Export genome browser viewable files}
\usage{
exportCombinedMultivariate(hmm, filename, what = c("combstates"),
  exclude.states = "", include.states = NULL, trackname = NULL,
  header = TRUE, separate.files = FALSE)
}
\arguments{
\item{hmm}{A \code{\link{chromstaR_combinedMultiHMM}} object or file that contains such an object.}

\item{filename}{The name of the file that will be written. The ending ".bed.gz" for combinatorial states will be appended. Any existing file will be overwritten.}

\item{what}{A character vector specifying what will be exported. Supported are \code{c('combstates')}.}

\item{exclude.states}{A vector of combinatorial states that will be excluded from export.}

\item{include.states}{A vector of combinatorial states that will be exported. If specified, \code{exclude.states} is ignored.}

\item{trackname}{Name that will be used in the "track name" field of the BED file.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for each condition.}
}
\description{
Export multivariate calls as genome browser viewable file
}
\details{
Export \code{\link{chromstaR_combinedMultiHMM}} objects as files which can be uploaded into a genome browser. Combinatorial states are exported in BED format (.bed.gz).
}
\examples{
### Multivariate peak calling ###
## Load example multivariate Hidden Markov Model
data(example.multi.HMM)
## Export the binned read counts and peaks for each track
\donttest{
exportMultivariate(example.multi.HMM, exclude.states=c(0),
                  filename='chromstaR_example.multi.HMM',
                  what=c('counts','peaks'))
}
}
\author{
Aaron Taudt
}

