% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export.R
\name{exportGRanges}
\alias{exportGRanges}
\title{Export genome browser viewable files}
\usage{
exportGRanges(gr, trackname, filename = "chromstaR_GRanges_regions",
  header = TRUE)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{trackname}{The name that will be used as track name and description in the header.}

\item{filename}{The name of the file that will be written. The ending ".bed.gz". Any existing file will be overwritten.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Export GRanges as genome browser viewable file
}
\details{
Export regions from \code{\link{GRanges}} as a file which can be uploaded into a genome browser. Regions are exported in BED format (.bed.gz).
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportMultivariate}}
}

