% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportGRanges}
\alias{exportGRanges}
\title{Export genome browser viewable files}
\usage{
exportGRanges(gr, trackname, filename = "chromstaR_GRanges_regions",
  header = TRUE, orderByScore = FALSE, append = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{trackname}{The name that will be used as track name and description in the header.}

\item{filename}{The name of the file that will be written. The ending ".bed.gz". Any existing file will be overwritten.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{orderByScore}{If set to \code{TRUE}, the lines in the output will be ordered by column 'score' in descending order.}

\item{append}{Whether or not to append to an existing file.}
}
\value{
\code{NULL}
}
\description{
Export GRanges as genome browser viewable file
}
\details{
Export regions from \code{\link{GRanges}} as a file which can be uploaded into a genome browser. Regions are exported in BED format (.bed.gz).
}
\examples{
### Export regions with read counts above 20 ###
# Get an example BED file with ChIP-seq reads
bedfile <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bed.gz",
                       package="chromstaRData")
# Bin the BED file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(bedfile, assembly=rn4_chrominfo, binsize=1000,
                  chromosomes='chr12')
plot(binned)
# Export regions with read count above 20
exportGRanges(binned[binned$counts > 20], filename=tempfile(),
             trackname='read counts above 20')

}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportMultivariate}}
}

