% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.R
\name{exportMultivariate}
\alias{exportMultivariate}
\title{Export genome browser viewable files}
\usage{
exportMultivariate(multi.hmm, filename, what = c("combstates", "peaks",
  "reads"), exclude.states = 0, include.states = NULL, header = TRUE,
  separate.files = FALSE)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} object or file that contains such an object.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either ".bed.gz" for combinatorial states and peak-calls or ".wig.gz" for read counts. Any existing file will be overwritten.}

\item{what}{A character vector specifying what will be exported. Supported are \code{c('combstates', 'peaks', 'reads')}.}

\item{exclude.states}{A vector of combinatorial states that will be excluded from export.}

\item{include.states}{A vector of combinatorial states that will be exported. If specified, \code{exclude.states} is ignored.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Export multivariate calls and read counts as genome browser viewable file
}
\details{
Export \code{\link{chromstaR_univariateHMM}} objects as files which can be uploaded into a genome browser. Combinatorial states and peak-calls are exported in BED format (.bed.gz) and read counts are exported in WIGGLE format (.wig.gz).
}
\examples{
### Multivariate peak calling ###
## Load example multivariate Hidden Markov Model
data(example.multi.HMM)
## Export the binned read counts and peaks for each track
\donttest{
exportMultivariate(example.multi.HMM, exclude.states=c(0),
                  filename='chromstaR_example.multi.HMM',
                  what=c('reads','peaks'))
}
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportBinnedData}}
}

