% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportMultivariate}
\alias{exportMultivariate}
\title{Export genome browser viewable files}
\usage{
exportMultivariate(multi.hmm, filename, what = c("combinations", "peaks",
  "counts"), exclude.states = "[]", include.states = NULL,
  trackname = NULL, header = TRUE, separate.files = FALSE,
  orderByScore = TRUE)
}
\arguments{
\item{multi.hmm}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either ".bed.gz" for combinatorial states and peak-calls or ".wig.gz" for read counts. Any existing file will be overwritten.}

\item{what}{A character vector specifying what will be exported. Supported are \code{c('combinations', 'peaks', 'counts')}.}

\item{exclude.states}{A character vector with combinatorial states that will be excluded from export.}

\item{include.states}{A character vector with combinatorial states that will be exported. If specified, \code{exclude.states} is ignored.}

\item{trackname}{Name that will be used in the "track name" field of the BED file.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for peaks if \code{what} contains 'peaks' or 'counts'.}

\item{orderByScore}{Logical indicating whether or not to sort entries in BED file by score.}
}
\description{
Export multivariate calls and read counts as genome browser viewable file
}
\details{
Export \code{\link{uniHMM}} objects as files which can be uploaded into a genome browser. Combinatorial states and peak-calls are exported in BED format (.bed.gz) and read counts are exported in WIGGLE format (.wig.gz).
}
\examples{
### Multivariate peak calling ###
## Load example multivariate Hidden Markov Model
data(example.multi.HMM)
## Export the binned read counts and peaks for each track
\donttest{
exportMultivariate(example.multi.HMM, exclude.states=c(0),
                  filename='chromstaR_example.multi.HMM',
                  what=c('counts','peaks'))
}
}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportUnivariates}}, \code{\link{exportBinnedData}}
}

