% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportUnivariates}
\alias{exportUnivariates}
\title{Export genome browser viewable files}
\usage{
exportUnivariates(hmm.list, filename, what = c("peaks", "counts"),
  header = TRUE, separate.files = FALSE)
}
\arguments{
\item{hmm.list}{A \code{list()} of \code{\link{uniHMM}} objects or vector of files that contain such objects.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either ".bed.gz" for peak-calls or ".wig.gz" for read counts. Any existing file will be overwritten.}

\item{what}{A character vector specifying what will be exported. Supported are \code{c('peaks', 'counts')}.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{separate.files}{A logical indicating whether or not to produce separate files for each hmm in \code{hmm.list}.}
}
\value{
\code{NULL}
}
\description{
Export univariate peak-calls and read counts as genome browser viewable file
}
\details{
Export \code{\link{uniHMM}} objects as files which can be uploaded into a genome browser. Peak-calls are exported in BED format (.bed.gz) and read counts are exported in WIGGLE format (.wig.gz).
}
\examples{
## Get an example BED file
bedfile <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bed.gz",
                       package="chromstaRData")
## Bin the BED file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(bedfile, assembly=rn4_chrominfo, binsize=1000,
                  chromosomes='chr12')
## Fit the univariate Hidden Markov Model
hmm <- callPeaksUnivariate(binned, max.time=60, eps=1)
## Export
exportUnivariates(list(hmm), filename=tempfile(), what=c('peaks','counts'))

}
\author{
Aaron Taudt
}
\seealso{
\code{\link{exportBinnedData}}, \code{\link{exportMultivariate}}
}

