% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{expressionAtPercentageOverlap}
\alias{expressionAtPercentageOverlap}
\title{Mean expression at percentage overlap}
\usage{
expressionAtPercentageOverlap(multi.hmm, expression)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} or a file that contains such an object.}

\item{expression}{A \code{\link{GRanges}} object with metadata column 'expression', containing the expression value for each range of the feature.}
}
\value{
A list with vectors of mean expression values per percentile for each combinatorial state.
}
\description{
Get the average expression for each percentage of overlap of combinatorial state with feature.
}
\author{
Aaron Taudt
}

