% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{expressionOverlap}
\alias{expressionOverlap}
\title{Overlap with expression data}
\usage{
expressionOverlap(multi.hmm, expression)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} or a file that contains such an object.}

\item{expression}{A \code{\link{GRanges}} object with metadata column 'expression', containing the expression value for each range.}
}
\value{
A named list with expression values.
}
\description{
Get the expression values that overlap with each combinatorial state.
}
\author{
Aaron Taudt
}

