% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getStateColors}
\alias{getStateColors}
\title{Get state colors}
\usage{
getStateColors(labels = NULL)
}
\arguments{
\item{labels}{Any combination of \code{c("zero-inflation","unmodified","modified", "total", "counts")}.}
}
\value{
A character vector with colors.
}
\description{
Get the colors that are used for plotting.
}
\examples{
cols <- getStateColors()
pie(1:length(cols), col=cols, labels=names(cols))
}
\seealso{
\code{\link{plotting}}
}
