% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapCombinations}
\alias{heatmapCombinations}
\title{Plot a heatmap of combinatorial states}
\usage{
heatmapCombinations(model = NULL, marks = NULL)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{marks}{A character vector with histone marks. If specified, \code{model} will be ignored.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a heatmap that shows the binary presence/absence of marks for the different combinations.
}
\examples{
marks <- c('H3K4me3','H3K27me3','H4K20me1')
heatmapCombinations(marks=marks)

file <- system.file("data","multivariate_mode-mark_condition-SHR.RData",
                    package="chromstaR")
heatmapCombinations(file)

}
\author{
Aaron Taudt
}
\seealso{
\code{\link{plotting}}
}

