% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapCountCorrelation}
\alias{heatmapCountCorrelation}
\title{Read count correlation heatmap}
\usage{
heatmapCountCorrelation(model, cluster = TRUE)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} or \code{\link{combinedMultiHMM}} object or file that contains such an object.}

\item{cluster}{Logical indicating whether or not to cluster the heatmap.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Heatmap of read count correlations (see \code{\link[stats]{cor}}).
}
\examples{
## Get an example multiHMM ##
file <- system.file("data","multivariate_mode-combinatorial_condition-SHR.RData",
                    package="chromstaR")
model <- get(load(file))
## Plot count correlations as heatmap
heatmapCountCorrelation(model)

}
\seealso{
\code{\link{plotting}}
}

