% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromFiles.R
\name{loadHmmsFromFiles}
\alias{loadHmmsFromFiles}
\title{Load HMMs from files}
\usage{
loadHmmsFromFiles(hmm.list, strict = FALSE)
}
\arguments{
\item{hmm.list}{A list of files that contain \code{\link{chromstaR_univariateHMM}} objects.}

\item{strict}{If any of the loaded objects is not a \code{\link{chromstaR_univariateHMM}} object, an error (\code{strict=TRUE}) or a warning (\code{strict=FALSE}) will be generated.}
}
\value{
A \code{list()} containing all loaded \code{\link{chromstaR_univariateHMM}} objects.
}
\description{
Load \code{\link{chromstaR_univariateHMM}} objects from file into a list.
}
\examples{
\donttest{
## Not run:
files.with.hmms <- list.files(folder.with.hmms, full=TRUE,
                             pattern='binsize_1000')
uni.HMMs <- loadHmmsFromFiles(files.with.hmms)
## End(Not run)
}
}
\author{
Aaron Taudt
}
\seealso{
loadMultiHmmsFromFiles
}

