% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHmmsFromFiles.R
\name{loadHmmsFromFiles}
\alias{loadHmmsFromFiles}
\title{Load \pkg{chromstaR} objects from file}
\usage{
loadHmmsFromFiles(files, check.class = c("GRanges", "uniHMM", "multiHMM",
  "combinedMultiHMM"))
}
\arguments{
\item{files}{A list of \code{\link{chromstaR-objects}} or a vector of files that contain such objects.}

\item{check.class}{Any combination of \code{c('GRanges', 'uniHMM', 'multiHMM', 'combinedMultiHMM')}. If any of the loaded objects does not belong to the specified class, an error is thrown.}
}
\value{
A list of \code{\link[chromstaR:chromstaR-objects]{chromstaR-object}}.
}
\description{
Wrapper to load \pkg{\link{chromstaR}} objects from file and check the class of the loaded objects.
}
\examples{
## Get an example BAM file
file <- system.file("extdata", "euratrans",
                      "lv-H3K27me3-BN-male-bio2-tech1.bam",
                       package="chromstaRData")
## Bin the file into bin size 1000bp
data(rn4_chrominfo)
binned <- binReads(file, assembly=rn4_chrominfo, binsizes=1000,
                  chromosomes='chr12')
## Fit the univariate Hidden Markov Model
hmm <- callPeaksUnivariate(binned, max.time=60, eps=1)
temp.file <- tempfile()
save(hmm, file=temp.file)
loaded.hmm <- loadHmmsFromFiles(temp.file)[[1]]
class(loaded.hmm)

}

