% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadFromFiles.R
\name{loadMultiHmmsFromFiles}
\alias{loadMultiHmmsFromFiles}
\title{Load multivariate HMMs from files}
\usage{
loadMultiHmmsFromFiles(hmm.list, strict = FALSE)
}
\arguments{
\item{hmm.list}{A list of files that contain \code{\link{chromstaR_multivariateHMM}} objects.}

\item{strict}{If any of the loaded objects is not a \code{\link{chromstaR_multivariateHMM}} object, an error (\code{strict=TRUE}) or a warning (\code{strict=FALSE}) will be generated.}
}
\value{
A list() containing all loaded \code{\link{chromstaR_multivariateHMM}} objects.
}
\description{
Load \code{\link{chromstaR_multivariateHMM}} objects from file into a list.
}
\examples{
\donttest{
## Not run:
files.with.hmms <- list.files(folder.with.hmms, full=TRUE,
                             pattern='binsize_1000')
multi.HMMs <- loadMultiHmmsFromFiles(files.with.hmms)
## End(Not run)
}
}
\author{
Aaron Taudt
}
\seealso{
loadHmmsFromFiles
}

