% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{multivariate_model}
\alias{multivariate_model}
\title{Multivariate HMM for demonstration purposes}
\format{A \code{\link{multiHMM}} object.}
\description{
A \code{\link{multiHMM}} object for demonstration purposes in examples of package \pkg{\link{chromstaR}}.
}
\examples{
## Get an example multiHMM
file <- system.file("data","multivariate_mode-combinatorial_condition-SHR.RData",
                    package="chromstaR")
model <- get(load(file))
}

