% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGenomeBrowser}
\alias{plotGenomeBrowser}
\title{Plot a genome browser view}
\usage{
plotGenomeBrowser(counts, peaklist = NULL, chr, start, end,
  countcol = "black", peakcols = NULL, style = "peaks",
  peakTrackHeight = 5)
}
\arguments{
\item{counts}{A \code{\link[GenomicRanges]{GRanges}} object with meta-data column 'counts'.}

\item{peaklist}{A named list() of \code{\link[GenomicRanges]{GRanges}} objects containing peak coordinates.}

\item{chr, start, end}{Chromosome, start and end coordinates for the plot.}

\item{countcol}{A character giving the color for the counts.}

\item{peakcols}{A character vector with colors for the peaks in \code{peaklist}.}

\item{style}{One of \code{c('peaks', 'density')}.}

\item{peakTrackHeight}{Relative height of the tracks given in \code{peaklist} compared to the \code{counts}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a simple genome browser view. This is useful for scripted genome browser snapshots.
}

