% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHistograms}
\alias{plotHistograms}
\title{Histograms of binned read counts with fitted mixture distribution}
\usage{
plotHistograms(model, ...)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{...}{Additional arguments (see \code{\link{plotHistogram}}).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot histograms of binned read counts with fitted mixture distributions from a \code{\link{multiHMM}} object.
}
\seealso{
\code{\link{plotting}}
}
