% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_plotting.R
\name{plotting}
\alias{plotting}
\title{chromstaR plotting functions}
\description{
This page provides an overview of all \pkg{\link{chromstaR}} plotting functions.
}
\details{
Plotting functions that work on \code{\link{uniHMM}} objects:
\describe{
  \item{\code{\link{plotHistogram}}}{}
  \item{\code{\link{plotKaryogram}}}{}
}
Plotting functions that work on \code{\link{multiHMM}} objects:
\describe{
  \item{\code{\link{plotHistograms}}}{}
  \item{\code{\link{heatmapCountCorrelation}}}{}
  \item{\code{\link{plotTransitionProbs}}}{}
  \item{\code{\link{heatmapCombinations}}}{}
}
Plotting functions that work on \code{\link{multiHMM}} and \code{\link{combinedMultiHMM}} objects:
\describe{
  \item{\code{\link{heatmapCountCorrelation}}}{}
  \item{\code{\link{plotEnrichCountHeatmap}}}{}
  \item{\code{\link{plotFoldEnrichment}}}{}
  \item{\code{\link{plotFoldEnrichHeatmap}}}{}
}
Other plotting functions:
\describe{
  \item{\code{\link{heatmapCombinations}}}{}
}
}

