% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateMultivariate}
\alias{simulateMultivariate}
\title{Simulate multivariate data}
\usage{
simulateMultivariate(bins, transition, emissions, weights, correlationMatrices,
  combstates, IDs, fragLen = 50)
}
\arguments{
\item{bins}{A \code{\link[GenomicRanges]{GRanges}} object for which reads will be simulated.}

\item{transition}{A matrix with transition probabilities.}

\item{emissions}{A list() with data.frames with emission distributions (see \code{\link{uniHMM}} entry 'distributions').}

\item{weights}{A list() with weights for the three univariate states.}

\item{correlationMatrices}{A list with correlation matrices.}

\item{combstates}{A vector with combinatorial states.}

\item{IDs}{A character vector with IDs.}

\item{fragLen}{Length of the simulated read fragments.}
}
\value{
A \code{list()} with entries $bins containing the simulated states and read count, $reads with simulated read coordinates.
}
\description{
Simulate known states, read counts and read coordinates using a multivariate Hidden Markov Model.
}

