% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeDiffPeaks}
\alias{summarizeDiffPeaks}
\title{Summary information about differential peak calls}
\usage{
summarizeDiffPeaks(multi.hmm)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} object.}
}
\value{
A \code{list()} object with various entries.
}
\description{
Get information about the differential peak calls in a \code{\link{chromstaR_multivariateHMM}} object.
}
\examples{
## Load example multivariate Hidden Markov Model
data(example.multi.HMM)
## Summary information about differentially modified peaks
summarizeDiffPeaks(example.multi.HMM)
}
\author{
Aaron Taudt
}

