% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizePeaks}
\alias{summarizePeaks}
\title{Summary information about called peaks}
\usage{
summarizePeaks(multi.hmm)
}
\arguments{
\item{multi.hmm}{A \code{\link{chromstaR_multivariateHMM}} object.}
}
\value{
A \code{list()} object with various entries.
}
\description{
Get information about the peak calls in a \code{\link{chromstaR_multivariateHMM}} object.
}
\examples{
## Load example multivariate Hidden Markov Model
data(example.multi.HMM)
## Summary information about peaks
summarizePeaks(example.multi.HMM)
}
\author{
Aaron Taudt
}

