% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCombinatorial.R
\name{transitionFrequencies}
\alias{transitionFrequencies}
\title{Transition frequencies of combinatorial states}
\usage{
transitionFrequencies(multi.hmms = NULL, zero.states = "[]",
  combstates = NULL)
}
\arguments{
\item{multi.hmms}{A list with \code{\link{multiHMM}} objects or a vector with filenames that contain such objects.}

\item{zero.states}{The string(s) which identifies the zero.states.}

\item{combstates}{Alternative input instead of \code{multi.hmms}: A list of combinatorial state vectors instead of HMMs. If this is specified, \code{multi.hmms} will be ignored.}
}
\value{
A data.frame with transition frequencies.
}
\description{
Get a table of transition frequencies between combinatorial states of different \code{\link{multiHMM}}s.
}
\author{
Aaron Taudt
}

