% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyAnalysis.R
\name{transitionFrequencies}
\alias{transitionFrequencies}
\title{Transition frequencies of combinatorial states}
\usage{
transitionFrequencies(multi.hmms = NULL, combined.hmm = NULL,
  zero.states = "[]", combstates = NULL)
}
\arguments{
\item{multi.hmms}{A named list with \code{\link{multiHMM}} objects or a vector with filenames that contain such objects.}

\item{combined.hmm}{A \code{\link{combinedMultiHMM}} object. If specified, \code{multi.hmms} is ignored.}

\item{zero.states}{The string(s) which identifies the zero.states.}

\item{combstates}{Alternative input instead of \code{multi.hmms}: A named list of combinatorial state vectors instead of HMMs. Names must be of the form "combination.X", where X is an arbitrary string. If this is specified, \code{multi.hmms} and \code{combined.hmm} will be ignored.}
}
\value{
A data.frame with transition frequencies.
}
\description{
Get a table of transition frequencies between combinatorial states of different \code{\link{multiHMM}}s.
}
\examples{
#=== Step 1: Preparation ===
## Prepare the file paths. Exchange this with your input and output directories.
inputfolder <- system.file("extdata","euratrans", package="chromstaRData")
outputfolder <- file.path(tempdir(), 'SHR-BN-example')
## Define experiment structure
data(experiment_table)
print(experiment_table)
## Define assembly
# This is only necessary if you have BED files, BAM files are handled automatically.
# For common assemblies you can also specify them as 'hg19' for example.
data(rn4_chrominfo)
head(rn4_chrominfo)

#=== Step 2: Run Chromstar ===
## Run ChromstaR
Chromstar(inputfolder, experiment.table=experiment_table,
         outputfolder=outputfolder, numCPU=2, binsize=1000, assembly=rn4_chrominfo,
         prefit.on.chr='chr12', chromosomes='chr12', mode='combinatorial', eps.univariate=1,
         eps.multivariate=1)
## Results are stored in 'outputfolder' and can be loaded for further processing
list.files(outputfolder)
model <- get(load(file.path(outputfolder,'combined', 'combined_mode-combinatorial.RData')))

#=== Step 3: Analysis ===
# Get frequencies
freqs <- transitionFrequencies(combined.hmm=model)
freqs$table

}
\author{
Aaron Taudt
}

